+++ src/v1p1beta1/asset_service_client.ts
@@ -48,6 +48,7 @@ const version = require('../../../package.json').version;
 export class AssetServiceClient {
   private _terminated = false;
   private _opts: ClientOptions;
+  private _providedCustomServicePath: boolean;
   private _gaxModule: typeof gax | typeof gax.fallback;
   private _gaxGrpc: gax.GrpcClient | gax.fallback.GrpcClient;
   private _protos: {};
@@ -59,6 +60,7 @@ export class AssetServiceClient {
     longrunning: {},
     batching: {},
   };
+  warn: (code: string, message: string, warnType?: string) => void;
   innerApiCalls: {[name: string]: Function};
   assetServiceStub?: Promise<{[name: string]: Function}>;
 
@@ -101,6 +103,9 @@ export class AssetServiceClient {
     const staticMembers = this.constructor as typeof AssetServiceClient;
     const servicePath =
       opts?.servicePath || opts?.apiEndpoint || staticMembers.servicePath;
+    this._providedCustomServicePath = !!(
+      opts?.servicePath || opts?.apiEndpoint
+    );
     const port = opts?.port || staticMembers.port;
     const clientConfig = opts?.clientConfig ?? {};
     const fallback =
@@ -176,6 +181,9 @@ export class AssetServiceClient {
     // of calling the API is handled in `google-gax`, with this code
     // merely providing the destination and request information.
     this.innerApiCalls = {};
+
+    // Add a warn function to the client constructor so it can be easily tested.
+    this.warn = gax.warn;
   }
 
   /**
@@ -204,7 +212,8 @@ export class AssetServiceClient {
           )
         : // eslint-disable-next-line @typescript-eslint/no-explicit-any
           (this._protos as any).google.cloud.asset.v1p1beta1.AssetService,
-      this._opts
+      this._opts,
+      this._providedCustomServicePath
     ) as Promise<{[method: string]: Function}>;
 
     // Iterate over each of the methods that the service provides
@@ -296,7 +305,7 @@ export class AssetServiceClient {
   // -------------------
 
   searchAllResources(
-    request: protos.google.cloud.asset.v1p1beta1.ISearchAllResourcesRequest,
+    request?: protos.google.cloud.asset.v1p1beta1.ISearchAllResourcesRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -377,7 +386,7 @@ export class AssetServiceClient {
    *   for more details and examples.
    */
   searchAllResources(
-    request: protos.google.cloud.asset.v1p1beta1.ISearchAllResourcesRequest,
+    request?: protos.google.cloud.asset.v1p1beta1.ISearchAllResourcesRequest,
     optionsOrCallback?:
       | CallOptions
       | PaginationCallback<
@@ -555,7 +564,7 @@ export class AssetServiceClient {
     ) as AsyncIterable<protos.google.cloud.asset.v1p1beta1.IStandardResourceMetadata>;
   }
   searchAllIamPolicies(
-    request: protos.google.cloud.asset.v1p1beta1.ISearchAllIamPoliciesRequest,
+    request?: protos.google.cloud.asset.v1p1beta1.ISearchAllIamPoliciesRequest,
     options?: CallOptions
   ): Promise<
     [
@@ -631,7 +640,7 @@ export class AssetServiceClient {
    *   for more details and examples.
    */
   searchAllIamPolicies(
-    request: protos.google.cloud.asset.v1p1beta1.ISearchAllIamPoliciesRequest,
+    request?: protos.google.cloud.asset.v1p1beta1.ISearchAllIamPoliciesRequest,
     optionsOrCallback?:
       | CallOptions
       | PaginationCallback<